requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
celery[redis]
flower
celery_singleton